pyproject.toml
[tool.poetry]
name = "newsletter_gen"
version = "0.1.0"
description = "newsletter_gen using crewAI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = { extras = ["tools"], version = "^0.35.8" }
crewai-tools = "^0.4.6"
exa-py = "^1.0.12"
streamlit = "^1.36.0"
python-dotenv = "^1.0.1"
PyPDF2 = "^3.0.1"
python-docx = "^1.1.0"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
youtube-transcript-api = "^0.6.2"

[tool.poetry.scripts]
newsletter_gen = "newsletter_gen.main:run"
train = "newsletter_gen.main:train"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"